Create CAPSTONEPROJECT
After reviewing the literature, most of the research was positive for both the student and teacher in the utilization of SDSRS in the classroom.  The use of SDSRS in the classroom improving participation was a central theme throughout the research.  The use of SDSRS does promote active learning.  Teachers and students liked the immediate feedback feature and students especially liked the anonymity that the SDSRS provided